Fix -Wextra warnings: virtual dtors, deprecated copy, sign-compare
- Add virtual destructors to CallLegCreator, SimpleRelayCreator, and
  RegCacheStorageHandler (fixes undefined behavior when deleting
  derived objects through base unique_ptr)
- Add defaulted copy assignment operators to SdpPayload, SdpOrigin,
  SdpAttribute, AmSdp, PayloadMask, and trans_ticket (implicit copy
  assignment with user-provided copy ctor is deprecated in C++11 and
  will be removed in C++23)
- Fix signed/unsigned comparison warnings in ilbc.c, tcp_comm.c,
  avp.c, and test_rfc3261_musts.cpp

+++ apps/diameter_client/lib_dbase/avp.c
@@ -443,7 +443,7 @@ char*  AAAConvertAVPToString(AAA_AVP *avp, char *dest, unsigned int destLen)
       WARN("WARNING:AAAConvertAVPToString: don't know how to print"
 	   " this data type [%d] -> tryng hexa\n",avp->type);
     case AAA_AVP_DATA_TYPE:
-      for (i=0;i<avp->data.len&&l<destLen-1;i++)
+      for (i=0;i<avp->data.len&&(unsigned int)l<destLen-1;i++)
 	l+=snprintf(dest+l,destLen-l-1,"%x",
 		    ((unsigned char*)avp->data.s)[i]);
     }
